fix: wrong mapping an index to the length of array
Resolves #2957 

When drawing bars and trying to find `backgroundBarColor` for the current value, if the index `j` is greater than the given `backgroundBarColors` array, it is being reduced by the length of the array. However, this is not sufficient to map a value to the range of an array (think of an example of j being 4, and the array containing a single value). The mod of the index should be calculated and updated.